refactored to single component
