+++ packages/web-pkg/src/composables/actions/files/index.ts
@@ -18,7 +18,6 @@ export * from './useFileActionsRestore'
 export * from './useFileActionsSetImage'
 export * from './useFileActionsShowActions'
 export * from './useFileActionsShowDetails'
-export * from './useFileActionsShowEditTags'
 export * from './useFileActionsShowShares'
 export * from './useFileActionsCreateSpaceFromResource'
 export * from './useFileActionsCreateNewFolder'
